Outsource widgets in own component for future work
