+++ components/widgets/components/OpenQuestionWidget.tsx
